* commit '662d85e326490709ce299d156d9cfdaf02fd120b':
  query() doesn't throw RuntimeException anymore when there are now emails found
  Now using ms for sleeping to prevent confusion. Also removed default minimal values.

+++ mail-connector/src/main/java/eu/tsystems/mms/tic/testframework/mailconnector/util/EmailQuery.java
@@ -41,7 +41,7 @@ public class EmailQuery {
     private static final String MAX_READ_TRIES_PROPERTY = "MAX_READ_TRIES";
 
     private int retryCount = Integer.parseInt(PropertyManager.getProperty(MAX_READ_TRIES_PROPERTY, "20"));
-    private int pauseMs = Integer.parseInt(PropertyManager.getProperty(POLLING_TIMER_SECONDS_PROPERTY, "10"))*1000;
+    private long pauseMs = Integer.parseInt(PropertyManager.getProperty(POLLING_TIMER_SECONDS_PROPERTY, "10"))*1000;
     private SearchTerm searchTerm;
     private Predicate<Email> filter;
     private String folderName;
@@ -55,11 +55,11 @@ public class EmailQuery {
         return this;
     }
 
-    public int getPauseMs() {
+    public long getPauseMs() {
         return pauseMs;
     }
 
-    public EmailQuery setPauseMs(int millis) {
+    public EmailQuery setPauseMs(long millis) {
         this.pauseMs = millis;
         return this;
     }
